moved methods and classes placed in __init__.py during previous refactoring files back into named modules
